create main part of project

+++ backend/interaction_with_db.py
@@ -1,8 +1,10 @@
-from sqlalchemy import create_engine, Column, Integer, SmallInteger, Time, String
+from sqlalchemy import create_engine, Column, Integer, SmallInteger, Time, String, DateTime
 from sqlalchemy.orm import declarative_base, sessionmaker, relationship
+from sqlalchemy import and_, or_, not_
 from sqlalchemy.sql import func
 from sqlalchemy import ForeignKey
 from dotenv import load_dotenv
+from datetime import datetime, timedelta
 import os
 
 load_dotenv()
@@ -20,6 +22,7 @@ class User(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(64), nullable=False)
     password = Column(String(64), nullable=False)
+
     bookings = relationship("Booking", back_populates="user")
 
 
@@ -29,8 +32,9 @@ class Booking(Base):
     id = Column(Integer, primary_key=True, index=True)
     id_place = Column(SmallInteger, nullable=False)
     id_user = Column(Integer, ForeignKey('users.id', ondelete="CASCADE"), nullable=False)
-    st_time = Column(Time, nullable=False)
-    en_time = Column(Time, nullable=False)
+    st_datetime = Column(DateTime, nullable=False)
+    en_datetime = Column(DateTime, nullable=False)
+    duration = Column(Integer)
 
     user = relationship("User", back_populates="bookings")
 
@@ -42,27 +46,52 @@ Base.metadata.create_all(bind=engine)
 # ---------------------- Основные функции ----------------------
 
 # ---------------------- CRUD для USERS
-def find_user(name=None, password=None):
-    session = SessionLocal()
-    user = None
-    if name and password:
-        user = session.query(User).filter(User.name == name,
-                                          User.password == password).first()
-    elif name:
-        user = session.query(User).filter(User.name == name).first()
-    elif password:
-        user = session.query(User).filter(User.password == password).first()
-
-    session.close()
-    return user
-
-
 def create_user(name, password):
     session = SessionLocal()
     new_user = User(name=name, password=password)
     session.add(new_user)
     session.commit()
     session.close()
+
+
+def get_user(name=None, password=None, id=None):
+    session = SessionLocal()
+    try:
+        if not (name or password or id):
+            return None
+
+        # добавляем все фильтры не равные None
+        filters = []
+        if name:
+            filters.append(User.name == name)
+        if password:
+            filters.append(User.password == password)
+        if id:
+            filters.append(User.id == id)
+        user = session.query(User).filter(*filters).first()
+        if not user: return None
+
+        if user.bookings:
+            user_bookings = []
+            for booking in user.bookings:
+                user_bookings.append({
+                    "id": booking.id,
+                    "id_place": booking.id_place,
+                    "st_datetime": booking.st_datetime.isoformat(timespec='minutes'),
+                    "en_datetime": booking.en_datetime.isoformat(timespec='minutes'),
+                    "duration": booking.duration
+                })
+        else:
+            user_bookings = None
+
+        return {
+            "id": user.id,
+            "name": user.name,
+            # "password": user.password, не будем так делать пожалуй ;)
+            "bookings": user_bookings
+        }
+    finally:
+        session.close()
 
 
 def get_all_users():
@@ -73,7 +102,7 @@ def get_all_users():
         users_list.append({
             "id": user.id,
             "name": user.name,
-            "password": user.password
+            # "password": user.password
         })
     session.close()
     return users_list
@@ -97,29 +126,67 @@ def update_user(id, update_dict):
 def delete_user(id):
     session = SessionLocal()
     user = session.query(User).filter(User.id == id).first()
+    if not user:
+        session.close()
+        return False
+
     if user:
         session.delete(user)
         session.commit()
         session.close()
         return True
-    session.close()
-    return False
+
 
 # ---------------------- CRUD для BOOKING
-def create_booking(id_user, id_place, st_time, en_time):
+def create_booking(id_user, id_place, datetime_str, duration_minutes=120):
     session = SessionLocal()
-    new_booking = Booking(
-        id_user=id_user,
-        id_place=id_place,
-        st_time=st_time,
-        en_time=en_time
-    )
-    session.add(new_booking)
-    session.commit()
-    session.close()
-    return new_booking
+    try:
+        st_datetime = datetime.fromisoformat(datetime_str)
+        en_datetime = st_datetime + timedelta(minutes=duration_minutes)
 
-def get_booking(booking_id):
+        new_booking = Booking(
+            id_user=id_user,
+            id_place=id_place,
+            st_datetime=st_datetime,
+            en_datetime=en_datetime,
+            duration=duration_minutes
+        )
+
+        session.add(new_booking)
+        session.commit()
+        return new_booking
+    except:
+        return None
+    finally:
+        session.close()
+
+
+
+def get_bookings_by_datetime_range(start_str, end_str):
+    session = SessionLocal()
+
+    start = datetime.fromisoformat(start_str)
+    end = datetime.fromisoformat(end_str)
+
+    bookings = session.query(Booking).filter(
+        and_(
+            not_(and_(Booking.st_datetime < start, Booking.en_datetime < start)),
+            not_(and_(Booking.st_datetime > end, Booking.en_datetime > end))
+        )
+    ).all() # берет только тех у кого хоть какая-то часть лежит внутри интервала
+    session.close()
+
+    return [{
+        "id": b.id,
+        "id_user": b.id_user,
+        "id_place": b.id_place,
+        "st_datetime": b.st_datetime.isoformat(),
+        "en_datetime": b.en_datetime.isoformat(),
+        "duration": b.duration
+    } for b in bookings]
+
+
+def get_booking_by_id(booking_id):
     session = SessionLocal()
     booking = session.query(Booking).filter(Booking.id == booking_id).first()
     session.close()
@@ -128,53 +195,80 @@ def get_booking(booking_id):
             "id": booking.id,
             "id_place": booking.id_place,
             "id_user": booking.id_user,
-            "st_time": booking.st_time.strftime("%H:%M"),
-            "en_time": booking.en_time.strftime("%H:%M")
+            "st_datetime": booking.st_datetime.isoformat(timespec='minutes'),
+            "en_datetime": booking.en_datetime.isoformat(timespec='minutes'),
+            "duration": booking.duration,
+            "user": {
+                "id": booking.user.id,
+                "name": booking.user.name
+            } if booking.user else None
         }
     return None
 
+
 def get_all_bookings():
     session = SessionLocal()
-    bookings = session.query(Booking).all()
+    try:
+        bookings = session.query(Booking).all()
+        return [{
+            "id": b.id,
+            "id_place": b.id_place,
+            "id_user": b.id_user,
+            "st_datetime": b.st_datetime.isoformat(timespec='minutes'),
+            "en_datetime": b.en_datetime.isoformat(timespec='minutes'),
+            "duration": b.duration
+        } for b in bookings]
+    finally:
+        session.close()
+
+
+def get_user_bookings(user_id):
+    session = SessionLocal()
+    bookings = session.query(Booking).filter(
+        Booking.id_user == user_id
+    ).order_by(Booking.st_datetime).all()
     session.close()
+
     return [{
         "id": b.id,
         "id_place": b.id_place,
         "id_user": b.id_user,
-        "st_time": b.st_time.strftime("%H:%M"),
-        "en_time": b.en_time.strftime("%H:%M")
+        "st_datetime": b.st_datetime.isoformat(timespec='minutes'),
+        "en_datetime": b.en_datetime.isoformat(timespec='minutes'),
+        "duration_minutes": b.duration
     } for b in bookings]
 
-def get_user_bookings(user_id):
-    session = SessionLocal()
-    bookings = session.query(Booking).filter(Booking.id_user == user_id).all()
-    session.close()
-    return [{
-        "id": b.id,
-        "id_place": b.id_place,
-        "st_time": b.st_time.strftime("%H:%M"),
-        "en_time": b.en_time.strftime("%H:%M")
-    } for b in bookings]
 
-def update_booking(booking_id, update_data):
+def update_booking(booking_id, data):
     session = SessionLocal()
     booking = session.query(Booking).filter(Booking.id == booking_id).first()
-    if booking:
-        if 'id_place' in update_data:
-            booking.id_place = update_data['id_place']
-        if 'st_time' in update_data:
-            booking.st_time = update_data['st_time']
-        if 'en_time' in update_data:
-            booking.en_time = update_data['en_time']
-        session.commit()
-    session.close()
-    return bool(booking)
+    if not booking:
+        return False
 
-def delete_booking(booking_id):
+    if 'id_place' in data:
+        booking.id_place = data['id_place']
+
+    if 'st_datetime' in data:
+        new_st = datetime.fromisoformat(data['st_datetime'])
+        booking.st_datetime = new_st
+        booking.en_datetime = new_st + timedelta(minutes=booking.duration)
+
+    if 'duration' in data:
+        booking.duration = data['duration']
+        booking.en_datetime = booking.st_datetime + timedelta(minutes=data['duration'])
+
+    session.commit()
+    session.close()
+    return True
+
+def delete_booking_by_id(booking_id):
     session = SessionLocal()
     booking = session.query(Booking).filter(Booking.id == booking_id).first()
-    if booking:
-        session.delete(booking)
-        session.commit()
+    if not booking:
+        session.close()
+        return False
+    session.delete(booking)
+    session.commit()
     session.close()
-    return bool(booking)
+    return True
+
